Manually append legacy blueprints
On older versions of ember-cli, before ember-cli-legacy-blueprints was added, this new setup of fetching the ancestral blueprint didn’t work.

We resolve this by manually adding the legacy blueprints path that ember-cli-coffeescript depends on to the list of paths to check. If it is already present, it will be deduplicated by ember-cli.